Improving the structure of the project files.
